✨ Implemented a single Callout component
