--- packages/design-system/src/components/N8nPanelCallout/index.ts
@@ -1,3 +1,0 @@
-import N8nPanelCallout from './PanelCallout.vue';
-
-export default N8nPanelCallout;
