+++ packages/editor-ui/src/plugins/components.ts
@@ -51,7 +51,6 @@ import {
 	N8nButton,
 	N8nElButton,
 	N8nCallout,
-	N8nPanelCallout,
 	N8nCard,
 	N8nIcon,
 	N8nIconButton,
@@ -93,7 +92,6 @@ Vue.use(N8nAvatar);
 Vue.component('n8n-button', N8nButton);
 Vue.component('el-button', N8nElButton);
 Vue.component('n8n-callout', N8nCallout);
-Vue.component('n8n-panel-callout', N8nPanelCallout);
 Vue.component('n8n-card', N8nCard);
 Vue.component('n8n-form-box', N8nFormBox);
 Vue.component('n8n-form-inputs', N8nFormInputs);
